+++ src/components/favBook/FavBook.js
@@ -5,6 +5,9 @@ import { Badge, Container } from 'react-bootstrap';
 import Card from 'react-bootstrap/Card';
 import { Messages } from 'primereact/messages';
 import { useMountEffect } from 'primereact/hooks';
+import Swal from 'sweetalert2';
+
+
 export default function FavBooks() {
     const [isBookDeleted, setIsBookDeleted] = useState(false);
     const [favBook, setFavBook] = useState([]);
@@ -34,34 +37,72 @@ export default function FavBooks() {
         await fetch(`${url}update/${state}/${id}`, {
             method: 'PUT',
             headers: {
-                "Content-Type": "application/json"
+              "Content-Type": "application/json"
             }
-        }).then((response) => {
+          }).then((response) => {
             if (response.status === 202) {
-                getFavBook();
-                alert("Updated sucessfully");
+              getFavBook();
+              Swal.fire({
+                icon: 'success',
+                text: 'Moved Successfully',
+              });
             }
-        }).catch((error) => {
+          }).catch((error) => {
             alert(error);
-        });
+          });
+
 
     }
 
     async function handleDelete(bookId) {
-        await fetch(`${url}delete/${bookId}`, {
-            method: 'DELETE',
-            headers: {
-                "Content-Type": "application/json"
+        Swal.fire({
+            title: 'Are you sure?',
+            // text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!',
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`${url}delete/${bookId}`, {
+                    method: 'DELETE',
+                    headers: {
+                        "Content-Type": "application/json"
+                    }
+                }).then((response) => {
+                    if (response.status === 204) {
+                        // getCurrentBook();
+                        Swal.fire({
+                            icon: 'success',
+                            text: 'Book deleted successfully',
+                        });
+                    }
+                }).catch((error) => {
+                    alert(error);
+                });
+            } else if (result.dismiss === Swal.DismissReason.cancel) {
+                Swal.fire({
+                    text: 'Your book is safe from deletion',
+                });
             }
-        }).then((response) => {
-            if (response.status === 204) {
-                getFavBook();
-                alert('Book deleted sucessfully');
-                setIsBookDeleted(true);
-            }
+        });
+    }
 
-        }).catch((error) => {
-            alert((error));
+    async function handleAddToCart(obj){
+        const storedItems = JSON.parse(localStorage.getItem("cartItems") || "[]");
+
+        const book = {
+            id: obj.id,
+            name: obj.title,
+            price: obj.price
+        };
+        storedItems.push(book);
+
+        localStorage.setItem("cartItems", JSON.stringify(storedItems));
+        Swal.fire({
+            icon: 'success',
+            text: 'Book added to cart successfully',
         });
     }
