+++ src/components/finished/Finished.js
@@ -5,6 +5,9 @@ import { Badge, Container } from 'react-bootstrap';
 import Card from 'react-bootstrap/Card';
 import { Messages } from 'primereact/messages';
 import { useMountEffect } from 'primereact/hooks';
+import Swal from 'sweetalert2';
+
+
 export default function Finished() {
     const [finishedBook, setFinished] = useState([]);
     const msgs = useRef(null);
@@ -36,11 +39,12 @@ export default function Finished() {
                 "Content-Type": "application/json"
             },
         }).then((response) => {
-
             if (response.status === 202) {
                 getFinishedBook();
-                alert("Updated sucessfully");
-
+                Swal.fire({
+                    icon: 'success',
+                    text: 'Book updated successfully',
+                });
             }
         }).catch((error) => {
             alert(error);
@@ -48,18 +52,54 @@ export default function Finished() {
     }
 
     async function handleDelete(bookId) {
-        await fetch(`${url}delete/${bookId}`, {
-            method: 'DELETE',
-            headers: {
-                "Content-Type": "application/json"
+        Swal.fire({
+            title: 'Are you sure?',
+            // text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!',
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`${url}delete/${bookId}`, {
+                    method: 'DELETE',
+                    headers: {
+                        "Content-Type": "application/json"
+                    }
+                }).then((response) => {
+                    if (response.status === 204) {
+                        getFinishedBook();
+                        Swal.fire({
+                            icon: 'success',
+                            text: 'Book deleted successfully',
+                        });
+                    }
+                }).catch((error) => {
+                    alert(error);
+                });
+            }else if (result.dismiss === Swal.DismissReason.cancel) {
+                Swal.fire({
+                    text: 'Your book is safe from deletion',
+                });
             }
-        }).then((response) => {
-            if (response.status === 204) {
-                getFinishedBook();
-                alert('Movie deleted sucessfully');
-            }
-        }).catch((error) => {
-            alert((error));
+        });
+    }
+
+    async function handleAddToCart(obj){
+        const storedItems = JSON.parse(localStorage.getItem("cartItems") || "[]");
+
+        const book = {
+            id: obj.id,
+            name: obj.title,
+            price: obj.price
+        };
+        storedItems.push(book);
+
+        localStorage.setItem("cartItems", JSON.stringify(storedItems));
+        Swal.fire({
+            icon: 'success',
+            text: 'Book added to cart successfully',
         });
     }
 
@@ -103,11 +143,9 @@ export default function Finished() {
                                     </Container>
                                 </Card.Text>
                                 <div className="container d-flex">
-
                                     <h6>
                                         <Badge className="text-wrap" bg="secondary">{obj?.category}</Badge>
                                     </h6>
-
                                 </div>
                                 <Container className="d-flex flex-wrap">
                                     <Button onClick={() => { updateState(obj.id, 'current') }} className="mb-3 mx-1 btn-sm" variant="primary">
